Refactor: common code from all Computer Vision Pipelines to a parent class; Add: Tests for common functionality
