+++ src/deepsparse/pipelines/__init__.py
@@ -16,3 +16,4 @@
 # limitations under the License.
 
 from .schema_helpers import *
+from .cv_pipeline import *
